==> global_networks/09_wholews_netviz.Rmd <==
---
title: "09a_prev20_watershed_netviz"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message=FALSE}
library(igraph)
library(GGally)
library(ggpubr)

library(ggnetwork)
library(intergraph)

library(unikn)
options(bitmapType='cairo')

```

### Goal: make a figure with all three networks

```{r}

# import metadata
fung_abun <- readRDS("../intermediates/global/fully_filtered_p20_fungal_otu_table_matched_up.rds")
bact_abun <- readRDS("../intermediates/global/fully_filtered_p20_bact_otu_table_matched_up.rds")
cross_data_names <- c(rownames(fung_abun),rownames(bact_abun))

# igraphs
cross_no_dummy <- readRDS("../intermediates/global/cross20_igraph.rds")
fung_no_dummy <- readRDS("../intermediates/global/fung20_igraph.rds")
bact_no_dummy <- readRDS("../intermediates/global/bact20_igraph.rds")



g <- cross_no_dummy

h = asNetwork(g)

set.seed(123)
g_test <- ggnetwork(h, layout = "fruchtermanreingold")
g_test$domain <- ifelse(grepl("fung", g_test$vertex.names), "Fungi", "Bacteria")
g_test$domain <- factor(g_test$domain, levels=c("Fungi", "Bacteria"))


c20_fig <- ggplot(g_test, aes(x = x, y = y, xend = xend, yend = yend)) +
  geom_edges(color = "grey20", linewidth = 0.1) +
  geom_nodes(aes(color = domain)) +
  theme_blank() +
  scale_color_manual(values=c("#1F78B4", "#A6CEE3")) + # "#A6CEE3" "#1F78B4"
  labs(color = "Domain", title = "Interkingdom")

#ggsave("../figures/global/crossdom_prev20_plot_fb.png", width = 9, height = 6)



# for fungi
net_fung20 <- asNetwork(fung_no_dummy)

set.seed(123)
f20_fig <- ggplot(ggnetwork(net_fung20, layout = "fruchtermanreingold"), aes(x = x, y = y, xend = xend, yend = yend)) +
  geom_edges(color = "grey20", linewidth = 0.1) +
  geom_nodes(color = "#1F78B4") +
  theme_blank() +
  #scale_color_brewer(palette = "Paired") + # "#A6CEE3" "#1F78B4"
  labs(title = "Fungi")

#ggsave("../figures/global/fungi_prev20_plot.png", plot = f20_fig, width = 7, height = 6)


# for bacteria
net_bact20 <- asNetwork(bact_no_dummy)

set.seed(123)
b20_fig <- ggplot(ggnetwork(net_bact20, layout = "fruchtermanreingold"), aes(x = x, y = y, xend = xend, yend = yend)) +
  geom_edges(color = "grey20", linewidth = 0.1) +
  geom_nodes(color = "#a6cee3") +
  theme_blank() +
  #scale_color_brewer(palette = "Paired") + # "#A6CEE3" "#1F78B4"
  labs(title = "Bacteria")


#ggsave("../figures/global/bacteria_prev20_plot.png", plot = b20_fig, width=7, height = 6)


```

```{r}
# put them together
library(patchwork)

#abc <- ggarrange(f20_fig, b20_fig, c20_fig, nrow = 1, common.legend = TRUE, legend = "bottom") + bgcolor("white")

library(patchwork)

plots <- f20_fig + b20_fig + c20_fig & theme(legend.position = "bottom")
plots <- plots + plot_layout(guides = "collect") + guides(color = guide_legend(override.aes = list(size=4))) + theme(legend.text = element_text(size=13))

ggsave("../figures/finals/all_prev20_watershed_networks.png", width = 18, height = 6)


```
